workshop2/gcs-mcp-server/requirements.txt
# gcs_mcp_server/requirements.txt
fastmcp
google-cloud-storage
fastapi
uvicorn


